Adding serviceManagement

+++ api/Models/ServiceManagement/Entities/Service.cs
@@ -8,7 +8,7 @@ public class Service : EntityBase<int>
     public required DateTime AvailableFrom { get; set; }
     public required DateTime AvailableTo { get; set; }
     public required TimeSpan Duration { get; set; }
-    public required float Price { get; set; }
+    public required decimal Price { get; set; }
     
     public List<Reservation> Reservations { get; set; }
     public List<ServiceAvailability> ServiceAvailability { get; set; }
